workers/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
nats-py==2.6.0
boto3==1.34.0
python-multipart==0.0.6
python-pptx==0.6.23
PyPDF2==3.0.1
PyMuPDF==1.23.8
Pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
httpx==0.25.2
tenacity==8.2.3
structlog==23.2.0
python-json-logger==2.0.7
asyncpg==0.29.0
